backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
uuid==1.30
firebase-admin==6.2.0
google-cloud-bigquery==3.11.4
google-cloud-storage==2.10.0
gunicorn==21.2.0
google-generativeai==0.3.2
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.37.0
webdriver-manager==4.0.2
